agents/post-agent/requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
linkedin-api>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
textstat>=0.7.3
apify-client>=1.4.0
